Update MacAddressRandomizer.ps1
In this updated version of the script, the $newMacAddress variable is initialized inside the while loop and is based on randomly generated values for each octet of the MAC address. The Get-Random cmdlet is used to generate random values between 0 and 255 (inclusive) for each octet, and the ToString("X2") method is used to convert each value to a two-digit hexadecimal string. The resulting values are concatenated with hyphens to form the new MAC address.

The script then checks if the new MAC address is different from the previous one by comparing it to the MacAddress property of the $adapter variable. If the new MAC address is different, it sets the new MAC address using the Set-NetAdapterAdvancedProperty cmdlet. If the new MAC address is the same as the previous one, the script does nothing.